requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.15.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
audioread==3.0.1
black==25.1.0
blinker==1.9.0
braceexpand==0.1.7
Brlapi==0.8.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
configobj==5.0.6
contourpy==1.3.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
cytoolz==1.0.1
datasets==3.3.2
dbus-python==1.2.18
decorator==5.2.1
defer==1.0.6
dill==0.3.8
Distance==0.1.3
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker-pycreds==0.4.0
docopt==0.6.2
editdistance==0.8.1
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
fiddle==0.3.0
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
g2p-en==2.1.0
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.70.0
httplib2==0.20.2
huggingface-hub==0.29.2
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
importlib-metadata==4.6.4
inflect==7.5.0
intervaltree==3.1.0
iotop==0.6
ipython==8.34.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jiwer==3.1.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaldi-python-io==1.2.2
kaldiio==2.18.1
keyboard==0.13.5
keyring==23.5.0
kiwisolver==1.4.8
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
Levenshtein==0.27.1
lhotse==1.29.0
libcst==1.6.0
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities==0.14.0
lilcom==1.8.0
llvmlite==0.44.0
loguru==0.7.3
louis==3.20.0
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.1.6
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.30.0
nemo-toolkit==2.1.0
netifaces==0.11.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==1.24.4
oauthlib==3.2.0
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.16.3
optuna==4.2.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.8.0
pillow==11.1.0
plac==1.4.3
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyarrow==19.0.1
PyAudio==0.2.14
pybind11==2.13.6
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pyloudnorm==0.1.1
pyparsing==3.2.1
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyreadline3==3.5.4
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-xapp==2.2.1
pytorch-lightning==2.5.0.post0
pytz==2025.1
pyxdg==0.27
PyYAML==6.0.2
RapidFuzz==3.12.2
referencing==0.36.2
regex==2024.11.6
reportlab==3.6.8
requests==2.32.3
resampy==0.4.3
rich==13.9.4
rpds-py==0.23.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
sacremoses==0.1.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
SecretStorage==3.3.1
sentencepiece==0.2.0
sentry-sdk==2.22.0
setproctitle==1.3.5
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
soundfile==0.13.1
sox==1.5.0
soxr==0.5.0.post1
SQLAlchemy==2.0.38
stack-data==0.6.3
streamlit==1.43.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminator==2.1.1
text-unidecode==1.3
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
torch==2.6.0
torchmetrics==1.6.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typeguard==4.4.2
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
unattended-upgrades==0.1
urllib3==2.3.0
wadllib==1.3.6
wandb==0.19.8
watchdog==6.0.0
wcwidth==0.2.13
webdataset==0.2.111
Werkzeug==3.1.3
wget==3.2
win32_setctime==1.2.0
wrapt==1.17.2
xdg==5
xkit==0.0.0
xxhash==3.5.0
yarl==1.18.3
zipp==1.0.0
